update report structure and attempt pdf output
