+++ backend/src/controllers/reports/projectRollup.js
@@ -1,66 +1,15 @@
 const useCommonComponents = require("../useCommonComponents/index");
 const useController = require("../useController/index");
 const model = require("@models/reports/projectRollup");
+const utils = require("./utils");
 const what = { single: "report", plural: "reports" };
 const controller = useController(model, what);
+const _ = require("lodash");
 
 // Template and data reading
-const fs = require("fs");
-const path = require("path");
 const cdogs = useCommonComponents("cdogs");
-const pdfConfig = { responseType: "arraybuffer" };
-
-/**
- * Reads a file and encodes it to the specified format
- *
- * @param   {string} path     The path of the file to read
- * @param   {string} encoding The format with which to encode the file contents
- * @returns {string}
- */
-const loadTemplate = async (path, encoding = "base64") => {
-  let data;
-  try {
-    data = await fs.readFileSync(path);
-    data = data.toString(encoding);
-  } catch (err) {
-    console.error(`
-      error: ${err},
-      path: ${path},
-      data: ${path},
-    `);
-  }
-  return data;
-};
-
-const getDocumentApiBody = async (
-  data,
-  templateFileName,
-  templateType = "docx",
-  reportName = "report",
-  //convertTo = "pdf"
-  convertTo = "docx"
-) => {
-  const templateContent = await loadTemplate(
-    path.resolve(__dirname, `../../../reports/${templateFileName}`)
-  );
-  return {
-    data: data,
-    formatters:
-      '{"formatMoney":"_function_formatMoney|function(data) { return data.toFixed(2); }"}',
-    options: {
-      cacheReport: true,
-      convertTo: convertTo,
-      overwrite: true,
-      reportName: reportName,
-    },
-    template: {
-      content: templateContent,
-      encodingType: "base64",
-      fileType: templateType,
-    },
-  };
-};
-
+const pdfConfig = utils.pdfConfig;
+const getDocumentApiBody = utils.getDocumentApiBody;
 
 /**
  * Get a Project rollup Report for a specific array of portfolio.
@@ -82,11 +31,13 @@ controller.getProjectStatusRollup = async (request, reply) => {
     };
 
     if (result.rollup.length > 0) {
-      result.rollup = { portfolios: groupByPortfolio(result.rollup) };
+      result.rollup = { portfolios: groupByPortfolio(result.rollup, "portfolio_name") };
     }
 
-    const body = await getDocumentApiBody(result, "PA_StatusPortfolioRollup_template.docx");
-    const pdf = await cdogs.api.post("/template/render", body, pdfConfig);
+    // const body = await getDocumentApiBody(result, "PA_StatusPortfolioRollup_template.docx");
+    //const pdf = await cdogs.api.post("/template/render", body, pdfConfig);
+    const testBody = await getDocumentApiBody(result, "test.docx");
+    const pdf = await cdogs.api.post("/template/render", testBody, pdfConfig);
 
     // Inject the pdf data into the request object
     request.data = pdf;
@@ -100,13 +51,16 @@ controller.getProjectStatusRollup = async (request, reply) => {
   } catch (err) {
     reply.code(500);
     console.error(`
+
+
+
     ERROR:
-    ERROR:
-    ERROR:
-    ERROR:
-    ERROR:
-    ERROR: ${err}
-    `)
+    ${err}
+
+
+
+
+    `);
     return { message: `There was a problem looking up this Project rollup Report.` };
   }
 };
@@ -114,34 +68,24 @@ controller.getProjectStatusRollup = async (request, reply) => {
 /**
  * Separates an array of projects into groups by their portfolio.
  *
+ * this moves each array of projects onto a "projects" key, and the portfolio name onto a "portfolio" key
+ * the Carbone.io engine needs to see objects nested in arrays, or arrays nested in objects, or else it can't iterate through.
+ * arrays inside of arrays, or objects inside of objects can't be iterated properly (at depth n+1)
+ *
  * @param   {any[]}   rows Array of projects ordered by portfolio.
+ * @param   {string}  prop string matches property name we want to group the rows by
  * @returns {any[][]}
  */
-const groupByPortfolio = (rows) => {
-  const portfolioGroup = [];
-
-  const projectsKeyedByPortfolioName = rows.reduce(
-    (acc, row) => {
-      const key = row.portfolio_name;
-      const curGroup = acc[key] ?? [];
-
-      // this returns an object where the key is the portfolio name
-      // and the value is an array of projects belonging to that portfolio name
-      return ({ ...acc, [key]: [...curGroup, row] });
-    },
-    {}
-  );
-
-  // this moves each array of projects onto a "projects" key, and the portfolio name onto a "portfolio" key
-  // the Carbone.io engine needs to see objects nested in arrays, or arrays nested in objects, or else it can't iterate through.
-  // arrays inside of arrays, or objects inside of objects can't be iterated properly (at depth n+1)
-  for (let property in projectsKeyedByPortfolioName) {
-    portfolioGroup.push({
-      "portfolio_name": property,
-      "projects": [...projectsKeyedByPortfolioName[property]]
-    });
-  }
+const groupByPortfolio = (rows, prop) => {
+  const rowsByPortfolio = _.groupBy(rows, prop);
+  const portfolioGroup = _.map(rowsByPortfolio, (value, key) => {
+    return {
+      portfolio_name: key,
+      projects: [...value],
+    };
+  });
 
   return portfolioGroup;
 };
+
 module.exports = controller;
