+++ backend/src/controllers/reports/utils.js
@@ -1,0 +1,61 @@
+const fs = require("fs");
+const path = require("path");
+
+// pdf configuration
+const pdfConfig = { responseType: "arraybuffer" };
+
+/**
+ * Reads a file and encodes it to the specified format
+ *
+ * @param   {string} path     The path of the file to read
+ * @param   {string} encoding The format with which to encode the file contents
+ * @returns {string}
+ */
+const loadTemplate = async (path, encoding = "base64") => {
+  let data;
+  try {
+    data = await fs.readFileSync(path);
+    data = data.toString(encoding);
+  } catch (err) {
+    console.error(`
+      error: ${err},
+      path: ${path},
+      data: ${data},
+    `);
+  }
+  return data;
+};
+
+const getDocumentApiBody = async (
+  data,
+  templateFileName,
+  templateType = "docx",
+  reportName = "report",
+  convertTo = "docx"
+) => {
+  const templateContent = await loadTemplate(
+    path.resolve(__dirname, `../../../reports/${templateFileName}`)
+  );
+  return {
+    data: data,
+    formatters:
+      '{"formatMoney":"_function_formatMoney|function(data) { return data.toFixed(2); }"}',
+    options: {
+      cacheReport: true,
+      convertTo: convertTo,
+      overwrite: true,
+      reportName: reportName,
+    },
+    template: {
+      content: templateContent,
+      encodingType: "base64",
+      fileType: templateType,
+    },
+  };
+};
+
+module.exports = {
+  pdfConfig,
+  loadTemplate,
+  getDocumentApiBody,
+};
